Ruleset: prevent false positives on Symfony polyfill code
When `PHPCompatibility(SymfonyPolyfillPHPxx)` is run over the code in the Symfony polyfill repos itself, it will detect some non-issues.

The code in the files is all wrapped within proper `defined()`, `version_compare()` and/or `function_exists()` conditions and will never be executed on incompatible PHP versions.

This simple change prevents these non-issues from being reported.

This fix does rely on people having installed the package in a directory called `polyfill-php##`.