CI/GH Action: fix failing build
For those polyfill repos which are still actively maintained (PHP 7.2 and higher), the minimum required PHP version has been raised to PHP 7.1.

The polyfill repos rely on the Composer version negotation to serve the correct package version for the PHP version on which the install is run.

With that in mind, we should rely on the same and presume that if the polyfills are installed at version 1.20 or higher, that their minimum supported PHP version is PHP 7.1.