Composer: fix the version retrieved for the PHP 7.2 polyfill & fix CI
Symfony dropped support for PHP < 7.2 in the Symfony polyfills 1.31 series.
This means that the `1.31` version of the PHP 7.2 polyfill no longer contains any code/polyfills as the minimum supported PHP version has become PHP 72, so the maximum allowed version of the `symfony/polyfill-php72` package - for our purposes (testing) - needs to be fixed to v `1.30`.
This is a similar change as was previously applied for the older polyfills via PR 21.

As for the Symfony polyfills for PHP 7.3 and higher: these now have a PHP 7.2 minimum supported PHP version, so CI needs to be adjusted to allow for PHP 7.2 syntaxes being used in those packages.
It also means we need a separate build to safeguard PHP 7.1 compatibility at v 1.30 of the Polyfills (before the version drop of PHP < 7.2).
This is a similar change as was previously applied after the Symfony PHP < 7.1 version drop via PR 24.