PHPCompatibilitySymfonyPolyfillPHP56: update ruleset for PHPCompatibility 10.0-alpha1
Looks like one function call would at the very least run into a warning: https://3v4l.org/v3mZE
Having said that, it would need a combination of Mbstring function overload being into effect + Mbstring not being available for this warning to be hit, which is probably why nobody ever reported it to Symfony.

All in all, not our concern, but a bug in the Polyfill, so let's just ignore it.