Fix links after upstream branch rename
The Symfony project renamed the `master` branch to `main` and now to `1.x`.

Aside from that a number of polyfills are no longer actively maintained, which means that on the `1.x` branch, the repo is now empty, so we should link to the last useful release.

Includes fixing two typos.