Polyfills 8.4/8.5: remove overzealous exclusion
I already annotated when I added these attribute related exclusions that they weren't targeted enough and ran the risk of blanket ignoring more than we should.

I've now decided to remove these as the excludes are too risky, meaning that end-users may end up missing legitimate PHP cross-version compatibility issues because of it.

This highlights that more needs to change in how attributes are detected and flagged in PHPCompatibility itself. Once that improvement is made, these tests should be re-enabled and more targeted excludes should be added.

Also see 74, which I have opened as a follow-up to this PR.